+++ src/core/background-attachment/index.test.ts
@@ -9,4 +9,3 @@ test('backgroundAttachment: keyword', () => {
     }
     expect(received).toEqual(expected)
 })
-