+++ src/core/background-attachment/index.test.ts
@@ -1,5 +1,4 @@
 import { style } from '../../index'
-import { styleCSS__ } from '../../integrations/jss'
 
 test('backgroundAttachment: keyword', () => {
     const received = style({
@@ -7,16 +6,6 @@ test('backgroundAttachment: keyword', () => {
     })
     const expected = {
         backgroundAttachment: 'fixed',
-    }
-    expect(received).toEqual(expected)
-})
-
-test('background-attachment: keyword', () => {
-    const received = styleCSS__({
-        'background-attachment': 'fixed',
-    })
-    const expected = {
-        'background-attachment': 'fixed',
     }
     expect(received).toEqual(expected)
 })