+++ src/core/font-variant-caps/index.ts
@@ -17,9 +17,9 @@ type FontVariantCapsPropValue =
 export type FontVariantCapsDeclaration = {
     /**
      * Maps to CSS's **`font-variant-caps`** property
-     * @category RBProperty
-     * @formalSyntaxForValue 	normal | small-caps | all-small-caps | petite-caps | all-petite-caps | unicase | titling-caps
-     * @implementationReference https://drafts.csswg.org/css-fonts-3/#propdef-font-variant-caps
+     * @initial normal
+     * @definition https://www.w3.org/TR/2020/WD-css-fonts-4-20201117/#font-variant-caps-prop
+     * @specification {@link https://www.w3.org/TR/2018/REC-css-fonts-3-20180920/ CSS Fonts Module Level 4}
      */
     fontVariantCaps: FontVariantCapsPropValue
 }