+++ src/core/font-weight/index.ts
@@ -2,7 +2,7 @@ import { GlobalCssKeyword, ValueOrFunc } from '../shared'
 
 export type FontWeightCSSProp = 'font-weight'
 
-type FontWeight = 'normal' | 'bold' | 'bolder' | 'lighter' | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900
+type FontWeight = 'normal' | 'bold' | 'bolder' | 'lighter' | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900 | 1000
 
 /** @hide */
 type FontWeightPropValue = FontWeight | GlobalCssKeyword
@@ -13,10 +13,9 @@ type FontWeightPropValue = FontWeight | GlobalCssKeyword
 export type FontWeightDeclaration = {
     /**
      * Maps to CSS's **`font-weight`** property
-     * @category RBProperty
-     * @formalSyntaxForValue normal | bold | bolder | lighter | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900
-     * @added 0.2.0
-     * @implementationReference https://www.w3.org/TR/2018/REC-css-fonts-3-20180920/#font-weight-prop
+     * @initial normal
+     * @definition https://www.w3.org/TR/2020/WD-css-fonts-4-20201117/#fo   nt-weight
+     * @specification {@link https://www.w3.org/TR/2018/REC-css-fonts-3-20180920/ CSS Fonts Module Level 4}
      */
     fontWeight: FontWeightPropValue
 }