+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './core'
-export * from './integrations'
 export { toNum } from './utils'