+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Chagngelog
 
+## [0.4.56]
+
+### CHANGED
+- Changed/fixed the docs for some properties.
+- Added the numerical value 1000 as an allowed value for fontWeight.
+
 ## [0.4.55]
 
 ## FIXED